+++ DarkCodex/changelog.md
@@ -11,6 +11,7 @@
 - made gather power ability visible
 - extra wild talent
 - option to allow achievements; this also clears the corresponding flag from future save files (otherwise uninstalling all mods would prevent achievements again)
+- impale infusion
 
 ### Changed
 - testing out new parameters on mobile gathering